ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
librosa>=0.10.2
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic==2.5.2
python-dotenv==1.0.0
soundfile
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
resampy
